grizzled-slf4j - Mark logger val as transient
Problem

When the Logging trait is mixed into a Serializable class
that class is unable to be serialized as the _logger
reference prevents serialization.

Solution

Mark the _logger reference in the Logging trait as
transient allowing for Serializable classes that mix in
the trait to be correctly serialized.

Result

Serializable classes which mix-in the Logging trait
can be serialized.